chore: address quality gate findings (#150)

+++ backend/src/main/java/org/booklore/service/metadata/extractor/Azw3MetadataExtractor.java
@@ -21,7 +21,7 @@ public class Azw3MetadataExtractor extends MobiBaseMetadataExtractor {
     protected void processFormatSpecificHeader(RandomAccessFile raf, PalmDBRecord record0, MobiHeader header, int headerLength) throws IOException {
         // For KF8 (AZW3), check for KF8 boundary and resource index
         if (header.isKF8 && headerLength >= 232) {
-            raf.seek(record0.offset + 16 + 192);
+            raf.seek((long) record0.offset + 16 + 192);
             long kf8Boundary = readInt(raf) & 0xFFFFFFFFL;
             if (kf8Boundary == KF8_BOUNDARY) {
                 log.debug("KF8 boundary found at expected location");
